Rename xpub.ts to xkeys.ts
